Move reminders to an own file
