+++ src/background/background.js
@@ -4,7 +4,8 @@ import browser from 'webextension-polyfill';
 import { canonicalizeUrl } from '../lib/url.js';
 import { valueConstantTicker } from '../lib/calltime.js';
 import { Database, Creator } from '../lib/db.js';
-import { getCurrentTab, openDashboardTab } from '../lib/tabs.js';
+import { getCurrentTab } from '../lib/tabs.js';
+import { initReminders } from '../lib/reminders.js';
 import _ from 'lodash';
 
 /**
@@ -39,21 +40,6 @@ async function rescheduleAlarm() {
   const pollTimer = valueConstantTicker(); // tracks time since last poll
   const tabTimers = {};
   const tabTitles = {};
-
-  // These variables are in milliseconds
-  const donationInterval = 1000 * 60 * 60 * 24 * 7; // One week
-  const toastInterval = 1000 * 60 * 60 * 24; // 24 hours
-  // Could use browser.runtime.connect for faster updates and less frequent
-  // polling but that feels like unnecessary complexity
-  // Don't use browser.storage.onChanged.addListener, it spams events for
-  // no reason
-  // Could also observe the db but there doesn't seem to be a light way to do
-  // that
-  // This variable is in minutes
-  const reminderCheckInterval = 5 / 60; // 5 seconds
-
-  // We'll toast when the browser starts
-  let lastToast = new Date(0);
 
   async function receiveCreator(msg) {
     console.log('receiveCreator: ' + JSON.stringify(msg));
@@ -154,61 +140,10 @@ error: ${JSON.stringify(message)}`
   browser.alarms.onAlarm.addListener(alarm => {
     if (alarm.name === 'heartbeat') {
       stethoscope();
-    } else if (alarm.name === 'reminderCheck') {
-      reminderCheck();
     }
   });
 
-  // This is outside vuex because we use it in background.js
-  async function isTimeToDonate() {
-    return db.getDonations(1).then(([lastDonation]) => {
-      // When the user hasn't donated yet, we won't know when to remind them
-      // to "donate again". But that might be nice, not telling them to
-      // donate until we know that they *can* donate.
-      const lastTime =
-        lastDonation === undefined ? new Date(0) : new Date(lastDonation.date);
-      return new Date() - lastTime > donationInterval;
-    });
-  }
-
-  function reminderCheck() {
-    isTimeToDonate()
-      .then(shouldDonate => {
-        if (shouldDonate) {
-          browser.browserAction.setBadgeBackgroundColor({
-            color: 'ForestGreen',
-          });
-          browser.browserAction.setBadgeText({ text: '🔔' });
-          browser.browserAction.setTitle({
-            title: 'Thankful: Reminder to donate',
-          });
-
-          if (new Date() - lastToast > toastInterval) {
-            // TODO: Add the logo to the notification
-            browser.notifications.create({
-              type: 'basic',
-              title: 'Thankful: Reminder to donate 🔔',
-              message:
-                "You haven't donated in a while, click here to go to the donation dashboard",
-            });
-            lastToast = new Date();
-          }
-        } else {
-          browser.browserAction.setBadgeText({ text: '' });
-          browser.browserAction.setTitle({ title: 'Thankful' });
-        }
-      })
-      .catch(err => {
-        console.error('Could not get reminderTimes:', err);
-      });
-  }
-
-  browser.notifications.onClicked.addListener(openDashboardTab);
-
-  reminderCheck();
-  browser.alarms.create('reminderCheck', {
-    periodInMinutes: reminderCheckInterval,
-  });
+  initReminders();
 
   stethoscope();
 })();
